Fix bugs in 32-bit back-porting

+++ sys/compat/linuxkpi/common/include/linux/interval_tree.h
@@ -8,9 +8,9 @@ struct interval_tree_node {
 		RB_ENTRY(interval_tree_node) rb_entry;
 		struct rb_node rb;
 	};
-	unsigned long long start;	/* Start of interval */
-	unsigned long long last;	/* Last location _in_ interval */
-	unsigned long long __subtree_last;
+	unsigned long start;	/* Start of interval */
+	unsigned long last;	/* Last location _in_ interval */
+	unsigned long __subtree_last;
 	unsigned long long it_magic;
 };
 
@@ -22,10 +22,10 @@ interval_tree_remove(struct interval_tree_node *node, struct rb_root *root);
 
 extern struct interval_tree_node *
 interval_tree_iter_first(struct rb_root *root,
-			 unsigned long long start, unsigned long long last);
+			 unsigned long start, unsigned long last);
 
 extern struct interval_tree_node *
 interval_tree_iter_next(struct interval_tree_node *node,
-			unsigned long long start, unsigned long long last);
+			unsigned long start, unsigned long last);
 
 #endif
